+++ packages/@webex/internal-plugin-metrics/src/metrics.types.ts
@@ -165,5 +165,7 @@ export type PreComputedLatencies =
   | 'internal.client.pageJMT'
   | 'internal.download.time'
   | 'internal.click.to.interstitial'
+  | 'internal.refresh.captcha.time'
+  | 'internal.get.u2c.time'
   | 'internal.call.init.join.req'
   | 'internal.other.app.api.time';