+++ packages/@webex/plugin-meetings/test/unit/spec/meeting/index.js
@@ -43,7 +43,7 @@ import {
   RemoteTrackType,
   MediaType,
 } from '@webex/internal-media-core';
-import {StreamEventNames} from '@webex/media-helpers';
+import {LocalStreamEventNames} from '@webex/media-helpers';
 import * as StatsAnalyzerModule from '@webex/plugin-meetings/src/statsAnalyzer';
 import EventsScope from '@webex/plugin-meetings/src/common/events/events-scope';
 import Meetings, {CONSTANTS} from '@webex/plugin-meetings';
@@ -99,7 +99,6 @@ import {
   MeetingInfoV2PolicyError,
 } from '../../../../src/meeting-info/meeting-info-v2';
 import {
-  CLIENT_ERROR_CODE_TO_ERROR_PAYLOAD,
   DTLS_HANDSHAKE_FAILED_CLIENT_CODE,
   ICE_FAILED_WITHOUT_TURN_TLS_CLIENT_CODE,
   ICE_FAILED_WITH_TURN_TLS_CLIENT_CODE,
@@ -613,36 +612,172 @@ describe('plugin-meetings', () => {
           assert.exists(meeting.joinWithMedia);
         });
 
-        describe('resolution', () => {
-          it('should success and return a promise', async () => {
-            meeting.join = sinon.stub().returns(Promise.resolve(test1));
-            meeting.addMedia = sinon.stub().returns(Promise.resolve(test4));
+        const fakeRoapMessage = {id: 'fake TURN discovery message'};
+        const fakeReachabilityResults = {id: 'fake reachability'};
+        const fakeTurnServerInfo = {id: 'fake turn info'};
+        const fakeJoinResult = {id: 'join result'};
 
-            const joinOptions = {correlationId: '12345'};
-            const mediaOptions = {audioEnabled: test1, allowMediaInLobby: true};
+        const joinOptions = {correlationId: '12345'};
+        const mediaOptions = {audioEnabled: true, allowMediaInLobby: true};
 
-            const result = await meeting.joinWithMedia({
-              joinOptions,
-              mediaOptions,
-            });
-            assert.calledOnceWithExactly(meeting.join, joinOptions);
-            assert.calledOnceWithExactly(meeting.addMedia, mediaOptions);
-            assert.deepEqual(result, {join: test1, media: test4});
-          });
+        let generateTurnDiscoveryRequestMessageStub;
+        let handleTurnDiscoveryHttpResponseStub;
+        let abortTurnDiscoveryStub;
+
+        beforeEach(() => {
+          meeting.join = sinon.stub().returns(Promise.resolve(fakeJoinResult));
+          meeting.addMedia = sinon.stub().returns(Promise.resolve(test4));
+
+          webex.meetings.reachability.getReachabilityResults.resolves(fakeReachabilityResults);
+
+          generateTurnDiscoveryRequestMessageStub = sinon
+            .stub(meeting.roap, 'generateTurnDiscoveryRequestMessage')
+            .resolves({roapMessage: fakeRoapMessage});
+          handleTurnDiscoveryHttpResponseStub = sinon
+            .stub(meeting.roap, 'handleTurnDiscoveryHttpResponse')
+            .resolves({turnServerInfo: fakeTurnServerInfo, turnDiscoverySkippedReason: undefined});
+          abortTurnDiscoveryStub = sinon.stub(meeting.roap, 'abortTurnDiscovery');
         });
 
-        describe('rejection', () => {
-          it('should error out and return a promise', async () => {
-            meeting.join = sinon.stub().returns(Promise.reject());
-            assert.isRejected(meeting.joinWithMedia({mediaOptions: {allowMediaInLobby: true}}));
+        it('should work as expected', async () => {
+          const result = await meeting.joinWithMedia({
+            joinOptions,
+            mediaOptions,
           });
 
-          it('should fail if called with allowMediaInLobby:false', async () => {
-            meeting.join = sinon.stub().returns(Promise.resolve(test1));
-            meeting.addMedia = sinon.stub().returns(Promise.resolve(test4));
-
-            assert.isRejected(meeting.joinWithMedia({mediaOptions: {allowMediaInLobby: false}}));
+          // check that TURN discovery is done with join and addMedia called
+          assert.calledOnceWithExactly(meeting.join, {
+            ...joinOptions,
+            roapMessage: fakeRoapMessage,
+            reachability: fakeReachabilityResults,
           });
+          assert.calledOnceWithExactly(generateTurnDiscoveryRequestMessageStub, meeting, true);
+          assert.calledOnceWithExactly(
+            handleTurnDiscoveryHttpResponseStub,
+            meeting,
+            fakeJoinResult
+          );
+          assert.calledOnceWithExactly(meeting.addMedia, mediaOptions, fakeTurnServerInfo);
+
+          assert.deepEqual(result, {join: fakeJoinResult, media: test4});
+        });
+
+        it("should not call handleTurnDiscoveryHttpResponse if we don't send a TURN discovery request with join", async () => {
+          generateTurnDiscoveryRequestMessageStub.resolves({roapMessage: undefined});
+
+          const result = await meeting.joinWithMedia({
+            joinOptions,
+            mediaOptions,
+          });
+
+          // check that TURN discovery is done with join and addMedia called
+          assert.calledOnceWithExactly(meeting.join, {
+            ...joinOptions,
+            roapMessage: undefined,
+            reachability: fakeReachabilityResults,
+          });
+          assert.calledOnceWithExactly(generateTurnDiscoveryRequestMessageStub, meeting, true);
+          assert.notCalled(handleTurnDiscoveryHttpResponseStub);
+          assert.notCalled(abortTurnDiscoveryStub);
+          assert.calledOnceWithExactly(meeting.addMedia, mediaOptions, undefined);
+
+          assert.deepEqual(result, {join: fakeJoinResult, media: test4});
+          assert.equal(meeting.turnServerUsed, false);
+        });
+
+        it('should call abortTurnDiscovery() if we do not get a TURN server info', async () => {
+          handleTurnDiscoveryHttpResponseStub.resolves({turnServerInfo: undefined, turnDiscoverySkippedReason: 'missing http response'});
+
+          const result = await meeting.joinWithMedia({
+            joinOptions,
+            mediaOptions,
+          });
+
+          // check that TURN discovery is done with join and addMedia called
+          assert.calledOnceWithExactly(meeting.join, {
+            ...joinOptions,
+            roapMessage: fakeRoapMessage,
+            reachability: fakeReachabilityResults,
+          });
+          assert.calledOnceWithExactly(generateTurnDiscoveryRequestMessageStub, meeting, true);
+          assert.calledOnceWithExactly(
+            handleTurnDiscoveryHttpResponseStub,
+            meeting,
+            fakeJoinResult
+          );
+          assert.calledOnceWithExactly(abortTurnDiscoveryStub);
+          assert.calledOnceWithExactly(meeting.addMedia, mediaOptions, undefined);
+
+          assert.deepEqual(result, {join: fakeJoinResult, media: test4});
+        });
+
+        it('should reject if join() fails', async () => {
+          const error = new Error('fake');
+          meeting.join = sinon.stub().returns(Promise.reject(error));
+          meeting.locusUrl = null; // when join fails, we end up with null locusUrl
+
+          await assert.isRejected(meeting.joinWithMedia({mediaOptions: {allowMediaInLobby: true}}));
+
+          assert.calledOnceWithExactly(abortTurnDiscoveryStub);
+
+          assert.calledWith(Metrics.sendBehavioralMetric,
+            BEHAVIORAL_METRICS.JOIN_WITH_MEDIA_FAILURE,
+            {
+              correlation_id: meeting.correlationId,
+              locus_id: undefined,
+              reason: error.message,
+              stack: error.stack,
+              leaveErrorReason: undefined,
+            },
+            {
+              type: error.name,
+            }
+          );
+        });
+
+        it('should fail if called with allowMediaInLobby:false', async () => {
+          meeting.join = sinon.stub().returns(Promise.resolve(test1));
+          meeting.addMedia = sinon.stub().returns(Promise.resolve(test4));
+
+          await assert.isRejected(
+            meeting.joinWithMedia({mediaOptions: {allowMediaInLobby: false}})
+          );
+        });
+
+        it('should call leave() if addMedia fails and ignore leave() failure', async () => {
+          const leaveError = new Error('leave error');
+          const addMediaError = new Error('fake addMedia error');
+
+          const leaveStub = sinon.stub(meeting, 'leave').rejects(leaveError);
+          meeting.addMedia = sinon.stub().rejects(addMediaError);
+
+          await assert.isRejected(
+            meeting.joinWithMedia({
+              joinOptions: {resourceId: 'some resource'},
+              mediaOptions: {allowMediaInLobby: true},
+            }),
+            addMediaError
+          );
+
+          assert.calledOnce(leaveStub);
+          assert.calledOnceWithExactly(leaveStub, {
+            resourceId: 'some resource',
+            reason: 'joinWithMedia failure',
+          });
+
+          assert.calledWith(Metrics.sendBehavioralMetric,
+            BEHAVIORAL_METRICS.JOIN_WITH_MEDIA_FAILURE,
+            {
+              correlation_id: meeting.correlationId,
+              locus_id: meeting.locusUrl.split('/').pop(),
+              reason: addMediaError.message,
+              stack: addMediaError.stack,
+              leaveErrorReason: leaveError.message,
+            },
+            {
+              type: addMediaError.name,
+            }
+          );
         });
       });
 
@@ -1412,7 +1547,6 @@ describe('plugin-meetings', () => {
       describe('#addMedia', () => {
         const muteStateStub = {
           handleClientRequest: sinon.stub().returns(Promise.resolve(true)),
-          applyClientStateLocally: sinon.stub().returns(Promise.resolve(true)),
         };
 
         let fakeMediaConnection;
@@ -2892,9 +3026,13 @@ describe('plugin-meetings', () => {
               getSettings: sinon.stub().returns({
                 deviceId: 'some device id',
               }),
-              muted: false,
+              userMuted: false,
+              systemMuted: false,
+              get muted() {
+                return this.userMuted || this.systemMuted;
+              },
               setUnmuteAllowed: sinon.stub(),
-              setMuted: sinon.stub(),
+              setUserMuted: sinon.stub(),
               setServerMuted: sinon.stub(),
               outputStream: {
                 getTracks: () => {
@@ -3129,28 +3267,50 @@ describe('plugin-meetings', () => {
                 if (stream !== undefined) {
                   switch (type) {
                     case 'audio':
-                      assert.calledOnceWithExactly(
-                        meeting.sendSlotManager.getSlot(MediaType.AudioMain).publishStream,
-                        stream
-                      );
+                      if (stream?.readyState === 'ended') {
+                        assert.notCalled(meeting.sendSlotManager.getSlot(MediaType.AudioMain).publishStream);
+                      } else {
+                        assert.calledOnceWithExactly(
+                          meeting.sendSlotManager.getSlot(MediaType.AudioMain).publishStream,
+                          stream
+                        );
+                      }
                       break;
                     case 'video':
-                      assert.calledOnceWithExactly(
-                        meeting.sendSlotManager.getSlot(MediaType.VideoMain).publishStream,
-                        stream
-                      );
+                      if (stream?.readyState === 'ended') {
+                        assert.notCalled(
+                          meeting.sendSlotManager.getSlot(MediaType.VideoMain).publishStream
+                        );
+                      } else {
+                        assert.calledOnceWithExactly(
+                          meeting.sendSlotManager.getSlot(MediaType.VideoMain).publishStream,
+                          stream
+                        );
+                      }
                       break;
                     case 'screenShareAudio':
-                      assert.calledOnceWithExactly(
-                        meeting.sendSlotManager.getSlot(MediaType.AudioSlides).publishStream,
-                        stream
-                      );
+                      if (stream?.readyState === 'ended') {
+                        assert.notCalled(
+                          meeting.sendSlotManager.getSlot(MediaType.AudioSlides).publishStream
+                        );
+                      } else {
+                        assert.calledOnceWithExactly(
+                          meeting.sendSlotManager.getSlot(MediaType.AudioSlides).publishStream,
+                          stream
+                        );
+                      }
                       break;
                     case 'screenShareVideo':
-                      assert.calledOnceWithExactly(
-                        meeting.sendSlotManager.getSlot(MediaType.VideoSlides).publishStream,
-                        stream
-                      );
+                      if (stream?.readyState === 'ended') {
+                        assert.notCalled(
+                          meeting.sendSlotManager.getSlot(MediaType.VideoSlides).publishStream
+                        );
+                      } else {
+                        assert.calledOnceWithExactly(
+                          meeting.sendSlotManager.getSlot(MediaType.VideoSlides).publishStream,
+                          stream
+                        );
+                      }
                       break;
                   }
                 }
@@ -3178,7 +3338,7 @@ describe('plugin-meetings', () => {
             }
           };
 
-          it('addMedia() works correctly when media is enabled without tracks to publish', async () => {
+          it('addMedia() works correctly when media is enabled without streams to publish', async () => {
             await meeting.addMedia();
             await simulateRoapOffer();
             await simulateRoapOk();
@@ -3244,8 +3404,8 @@ describe('plugin-meetings', () => {
             assert.calledTwice(locusMediaRequestStub);
           });
 
-          it('addMedia() works correctly when media is enabled with tracks to publish and track is muted', async () => {
-            fakeMicrophoneStream.muted = true;
+          it('addMedia() works correctly when media is enabled with streams to publish and stream is user muted', async () => {
+            fakeMicrophoneStream.userMuted = true;
 
             await meeting.addMedia({localStreams: {microphone: fakeMicrophoneStream}});
             await simulateRoapOffer();
@@ -3278,7 +3438,75 @@ describe('plugin-meetings', () => {
             assert.calledTwice(locusMediaRequestStub);
           });
 
-          it('addMedia() works correctly when media is disabled with tracks to publish', async () => {
+          it('addMedia() works correctly when media is enabled with tracks to publish and track is ended', async () => {
+            fakeMicrophoneStream.readyState = 'ended';
+
+            await meeting.addMedia({localStreams: {microphone: fakeMicrophoneStream}});
+            await simulateRoapOffer();
+            await simulateRoapOk();
+
+            // check RoapMediaConnection was created correctly
+            checkMediaConnectionCreated({
+              mediaConnectionConfig: expectedMediaConnectionConfig,
+              localStreams: {
+                audio: undefined,
+                video: undefined,
+                screenShareVideo: undefined,
+                screenShareAudio: undefined,
+              },
+              direction: {
+                audio: 'sendrecv',
+                video: 'sendrecv',
+                screenShare: 'recvonly',
+              },
+              remoteQualityLevel: 'HIGH',
+              expectedDebugId,
+              meetingId: meeting.id,
+            });
+            // and SDP offer was sent with the right audioMuted/videoMuted values
+            checkSdpOfferSent({audioMuted: true, videoMuted: true});
+            // check OK was sent with the right audioMuted/videoMuted values
+            checkOkSent({audioMuted: true, videoMuted: true});
+
+            // and that these were the only /media requests that were sent
+            assert.calledTwice(locusMediaRequestStub);
+          });
+
+          it('addMedia() works correctly when media is enabled with streams to publish and stream is system muted', async () => {
+            fakeMicrophoneStream.systemMuted = true;
+
+            await meeting.addMedia({localStreams: {microphone: fakeMicrophoneStream}});
+            await simulateRoapOffer();
+            await simulateRoapOk();
+
+            // check RoapMediaConnection was created correctly
+            checkMediaConnectionCreated({
+              mediaConnectionConfig: expectedMediaConnectionConfig,
+              localStreams: {
+                audio: fakeMicrophoneStream,
+                video: undefined,
+                screenShareVideo: undefined,
+                screenShareAudio: undefined,
+              },
+              direction: {
+                audio: 'sendrecv',
+                video: 'sendrecv',
+                screenShare: 'recvonly',
+              },
+              remoteQualityLevel: 'HIGH',
+              expectedDebugId,
+              meetingId: meeting.id,
+            });
+            // and SDP offer was sent with the right audioMuted/videoMuted values
+            checkSdpOfferSent({audioMuted: true, videoMuted: true});
+            // check OK was sent with the right audioMuted/videoMuted values
+            checkOkSent({audioMuted: true, videoMuted: true});
+
+            // and that these were the only /media requests that were sent
+            assert.calledTwice(locusMediaRequestStub);
+          });
+
+          it('addMedia() works correctly when media is disabled with streams to publish', async () => {
             await meeting.addMedia({
               localStreams: {microphone: fakeMicrophoneStream},
               audioEnabled: false,
@@ -3314,7 +3542,7 @@ describe('plugin-meetings', () => {
             assert.calledTwice(locusMediaRequestStub);
           });
 
-          it('addMedia() works correctly when media is disabled with no tracks to publish', async () => {
+          it('addMedia() works correctly when media is disabled with no streams to publish', async () => {
             await meeting.addMedia({audioEnabled: false});
             await simulateRoapOffer();
             await simulateRoapOk();
@@ -3347,7 +3575,7 @@ describe('plugin-meetings', () => {
             assert.calledTwice(locusMediaRequestStub);
           });
 
-          it('addMedia() works correctly when video is disabled with no tracks to publish', async () => {
+          it('addMedia() works correctly when video is disabled with no streams to publish', async () => {
             await meeting.addMedia({videoEnabled: false});
             await simulateRoapOffer();
             await simulateRoapOk();
@@ -3380,7 +3608,7 @@ describe('plugin-meetings', () => {
             assert.calledTwice(locusMediaRequestStub);
           });
 
-          it('addMedia() works correctly when screen share is disabled with no tracks to publish', async () => {
+          it('addMedia() works correctly when screen share is disabled with no streams to publish', async () => {
             await meeting.addMedia({shareAudioEnabled: false, shareVideoEnabled: false});
             await simulateRoapOffer();
             await simulateRoapOk();
@@ -3481,9 +3709,13 @@ describe('plugin-meetings', () => {
                 const fakeMicrophoneStream2 = {
                   on: sinon.stub(),
                   off: sinon.stub(),
-                  muted: false,
+                  userMuted: false,
+                  systemMuted: false,
+                  get muted() {
+                    return this.userMuted || this.systemMuted;
+                  },
                   setUnmuteAllowed: sinon.stub(),
-                  setMuted: sinon.stub(),
+                  setUserMuted: sinon.stub(),
                   outputStream: {
                     getTracks: () => {
                       return [
@@ -3720,12 +3952,12 @@ describe('plugin-meetings', () => {
           });
 
           [
-            {mute: true, title: 'muting a track before confluence is created'},
-            {mute: false, title: 'unmuting a track before confluence is created'},
+            {mute: true, title: 'user muting a track before confluence is created'},
+            {mute: false, title: 'user unmuting a track before confluence is created'},
           ].forEach(({mute, title}) =>
             it(title, async () => {
               // initialize the microphone mute state to opposite of what we do in the test
-              fakeMicrophoneStream.muted = !mute;
+              fakeMicrophoneStream.userMuted = !mute;
 
               await meeting.addMedia({localStreams: {microphone: fakeMicrophoneStream}});
               await stableState();
@@ -3734,10 +3966,54 @@ describe('plugin-meetings', () => {
 
               assert.equal(
                 fakeMicrophoneStream.on.getCall(0).args[0],
-                StreamEventNames.MuteStateChange
+                LocalStreamEventNames.UserMuteStateChange
               );
               const mutedListener = fakeMicrophoneStream.on.getCall(0).args[1];
               // simulate track being muted
+              fakeMicrophoneStream.userMuted = mute;
+              mutedListener(mute);
+
+              await stableState();
+
+              // nothing should happen
+              assert.notCalled(locusMediaRequestStub);
+              assert.notCalled(fakeRoapMediaConnection.update);
+
+              // now simulate roap offer and ok
+              await simulateRoapOffer();
+              await simulateRoapOk();
+
+              // it should be sent with the right mute status
+              checkSdpOfferSent({audioMuted: mute, videoMuted: true});
+              // check OK was sent with the right audioMuted/videoMuted values
+              checkOkSent({audioMuted: mute, videoMuted: true});
+
+              // nothing else should happen
+              assert.calledTwice(locusMediaRequestStub);
+              assert.notCalled(fakeRoapMediaConnection.update);
+            })
+          );
+
+          [
+            {mute: true, title: 'system muting a track before confluence is created'},
+            {mute: false, title: 'system unmuting a track before confluence is created'},
+          ].forEach(({mute, title}) =>
+            it(title, async () => {
+              // initialize the microphone mute state to opposite of what we do in the test
+              fakeMicrophoneStream.systemMuted = !mute;
+
+              await meeting.addMedia({localStreams: {microphone: fakeMicrophoneStream}});
+              await stableState();
+
+              resetHistory();
+
+              assert.equal(
+                fakeMicrophoneStream.on.getCall(0).args[0],
+                LocalStreamEventNames.UserMuteStateChange
+              );
+              const mutedListener = fakeMicrophoneStream.on.getCall(0).args[1];
+              // simulate track being muted
+              fakeMicrophoneStream.systemMuted = mute;
               mutedListener(mute);
 
               await stableState();
@@ -6138,6 +6414,65 @@ describe('plugin-meetings', () => {
             checkScreenShareVideoPublished(videoShareStream);
             checkScreenShareAudioPublished(audioShareStream);
           });
+
+          [
+            {
+              endedStream: 'microphone', 
+              streams: {
+                microphone: {
+                  readyState: 'ended',
+                },
+                camera: undefined,
+                screenShare: {
+                  audio: undefined,
+                  video: undefined,
+                },
+              },
+            },
+            {
+              endedStream: 'camera', 
+              streams: {
+                microphone: undefined,
+                camera: {
+                  readyState: 'ended',
+                },
+                screenShare: {
+                  audio: undefined,
+                  video: undefined,
+                },
+              },
+            },
+            {
+              endedStream: 'screenShare audio', 
+              streams: {
+                microphone: undefined,
+                camera: undefined,
+                screenShare: {
+                  audio: {
+                    readyState: 'ended',
+                  },
+                  video: undefined,
+                },
+              },
+            },
+            {
+              endedStream: 'screenShare video', 
+              streams: {
+                microphone: undefined,
+                camera: undefined,
+                screenShare: {
+                  audio: undefined,
+                  video: {
+                    readyState: 'ended',
+                  },
+                },
+              },
+            },
+          ].forEach(({endedStream, streams}) => {
+            it(`throws error if readyState of ${endedStream} is ended`, async () => {
+              assert.isRejected(meeting.publishStreams(streams));
+            })
+          });
         });
 
         describe('unpublishStreams', () => {
@@ -6272,11 +6607,11 @@ describe('plugin-meetings', () => {
         meeting.sendSlotManager.setNamedMediaGroups = sinon.stub().returns(undefined);
       });
       it('should throw error if not audio type', () => {
-        expect(() => meeting.setSendNamedMediaGroup(MediaType.VideoMain, 20)).to.throw(`cannot set send named media group which media type is ${MediaType.VideoMain}`)
-
+        expect(() => meeting.setSendNamedMediaGroup(MediaType.VideoMain, 20)).to.throw(
+          `cannot set send named media group which media type is ${MediaType.VideoMain}`
+        );
       });
       it('fails if there is no media connection', () => {
-
         meeting.mediaProperties.webrtcMediaConnection = undefined;
         meeting.setSendNamedMediaGroup('AUDIO-MAIN', 20);
         assert.notCalled(meeting.sendSlotManager.setNamedMediaGroups);
@@ -6285,8 +6620,10 @@ describe('plugin-meetings', () => {
       it('success if there is media connection', () => {
         meeting.isMultistream = true;
         meeting.mediaProperties.webrtcMediaConnection = true;
-        meeting.setSendNamedMediaGroup("AUDIO-MAIN", 20);
-        assert.calledOnceWithExactly(meeting.sendSlotManager.setNamedMediaGroups, "AUDIO-MAIN", [{type: 1, value: 20}]);
+        meeting.setSendNamedMediaGroup('AUDIO-MAIN', 20);
+        assert.calledOnceWithExactly(meeting.sendSlotManager.setNamedMediaGroups, 'AUDIO-MAIN', [
+          {type: 1, value: 20},
+        ]);
       });
     });
 
@@ -7360,6 +7697,7 @@ describe('plugin-meetings', () => {
         });
         it('listens to the self admitted guest event', (done) => {
           meeting.stopKeepAlive = sinon.stub();
+          meeting.updateLLMConnection = sinon.stub();
           meeting.locusInfo.emit({function: 'test', file: 'test'}, 'SELF_ADMITTED_GUEST', test1);
           assert.calledOnceWithExactly(meeting.stopKeepAlive);
           assert.calledThrice(TriggerProxy.trigger);
@@ -7370,6 +7708,7 @@ describe('plugin-meetings', () => {
             'meeting:self:guestAdmitted',
             {payload: test1}
           );
+          assert.calledOnce(meeting.updateLLMConnection);
           done();
         });
 
@@ -9495,6 +9834,13 @@ describe('plugin-meetings', () => {
               'https://board-a.wbx2.com/board/api/v1/channels/977a7330-54f4-11eb-b1ef-91f5eefc7bf3',
           };
 
+          const SHARE_TYPE = {
+            FILE:
+                'FILE',
+            DESKTOP:
+                'DESKTOP',
+          };
+
           const DEVICE_URL = {
             LOCAL_WEB: 'my-web-url',
             LOCAL_MAC: 'my-mac-url',
@@ -10492,6 +10838,36 @@ describe('plugin-meetings', () => {
 
               payloadTestHelper([data1, data2, data3]);
             });
+          });
+
+          describe('File Share  --> Desktop Share', () => {
+            it('Scenario #1: remote person A shares file then share desktop', () => {
+              const data1 = generateData(
+                  blankPayload,
+                  true,
+                  true,
+                  USER_IDS.ME,
+                  undefined,
+                  false,
+                  undefined,
+                  undefined,
+                  undefined,
+                  undefined,
+                  DEVICE_URL.LOCAL_WEB,
+                  SHARE_TYPE.FILE
+              );
+              const data2 = generateData(
+                  data1.payload,
+                  true,
+                  false,
+                  USER_IDS.ME,
+                  SHARE_TYPE.DESKTOP
+              );
+              const data3 = generateData(data2.payload, true, true, USER_IDS.ME);
+
+              payloadTestHelper([data1, data2, data3]);
+            });
+
           });
         });
       });